Refine Swift logging docs to better match other SDKs
