+++ examples/ios/Examples/Sync.swift
@@ -334,70 +334,6 @@ class Sync: AnonymouslyLoggedInTestCase {
         // :snippet-end:
     }
     
-    func testDefineCustomLogger() {
-        // :snippet-start: set-log-level-logger
-        // Access your app
-        let app = App(id: YOUR_APP_SERVICES_APP_ID)
-        
-        var logs: String = "" // :remove:
-        // Create an instance of `Logger` and define the log function to invoke.
-        let logger = Logger(level: .detail) { level, message in
-            logs += "\(Date.now) \(level) \(message) \n" // :remove:
-            "REALM DEBUG: \(Date.now) \(level) \(message) \n"
-        }
-        // :snippet-end:
-        Logger.shared = logger
-        XCTAssert(logs.isEmpty)
-        
-        let expectation = XCTestExpectation(description: "Populate some log entries")
-        app.login(credentials: Credentials.anonymous) { (result) in
-            switch result {
-            case .failure(let error):
-                fatalError("Login failed: \(error.localizedDescription)")
-            case .success(let user):
-                print("Login as \(user) successful")
-                XCTAssert(!logs.isEmpty)
-                expectation.fulfill()
-            }
-        }
-        wait(for: [expectation], timeout: 5)
-    }
-    
-    func testSetDefaultLogger() {
-        // :snippet-start: set-default-logger
-        // Access your app
-        let app = App(id: YOUR_APP_SERVICES_APP_ID)
-        
-        var logs: String = "" // :remove:
-        let logger = Logger(level: .info) { level, message in
-            logs += "\(Date.now) \(level) \(message) \n" // :remove:
-            "REALM DEBUG: \(Date.now) \(level) \(message) \n"
-        }
-        
-        // Set a logger as the default
-        Logger.shared = logger
-        XCTAssert(logs.isEmpty) // :remove:
-        
-        // After setting a default logger, you can change
-        // the log level at any point during the app lifecycle
-        Logger.shared.level = .debug
-        // :snippet-end:
-        XCTAssert(logs.isEmpty)
-        
-        let expectation = XCTestExpectation(description: "Populate some log entries")
-        app.login(credentials: Credentials.anonymous) { (result) in
-            switch result {
-            case .failure(let error):
-                fatalError("Login failed: \(error.localizedDescription)")
-            case .success(let user):
-                print("Login as \(user) successful")
-                XCTAssert(!logs.isEmpty)
-                expectation.fulfill()
-            }
-        }
-        wait(for: [expectation], timeout: 5)
-    }
-    
     // Skipping this test in CI because it fails when you run it with the other tests
     // with the error: `testSetCustomLogger(): std::logic_error: Cannot set the logger_factory after creating the sync client`
     // Running it by itself works.
