+++ source/sdk/swift.txt
@@ -22,6 +22,7 @@ Realm Swift SDK
    Actor-Isolated Realms </sdk/swift/actor-isolated-realm>
    Swift Concurrency </sdk/swift/swift-concurrency>
    Test and Debug </sdk/swift/test-and-debug>
+   Logging </sdk/swift/logging>
    Application Services </sdk/swift/application-services>
    Manage Users </sdk/swift/work-with-users>
    Sync Data </sdk/swift/sync>
